chore: rework nuvs export to use react-query
